Run the build steps sequentially
Building everything in parallel jumbles all the messages together,
making it difficult to read when something goes wrong. Sequence the
component build processes so that the next build step waits until the
prior one completes before beginning its own processing.

There is still a single build pipeline with the same structure. But now
the pipelines to build the device and companion components, and the
sub-pipelines to build each component bundle, are all lazy streams which
only begin processing when they are each read from for the first time.
And these pipelines are combined using `multistream`, which only begins
reading from the next stream once the previous stream ends. There is no
batching of intermediate build results: the component bundles flow all
the way to the end of the pipeline as soon as they are created.