Gate all compile options on component target
Change the interface of compile() to accept the component type, which
replaces `ecma`, `externals` and other build options which are dictated
by the component being built. The compile() function now derives all
those options from the component argument. This change also allows
for selectively enabling and disabling rollup plugins based on the
component. And since the `ecma` option is also needed for minification,
the output of compile() is also minified.

Image file imports are now only enabled for the settings component when
targeting SDK 3 and above. Importing an image is not going to work as
expected for the device or companion components, so it is disabled to
reduce the possibility of confusion.