Don't mutate native component ELF data in-place
The native component binary is loaded into memory to read the metadata,
then the raw binary data is passed along with the parsed metadata to
avoid having to read it a second time. `buf.swap64()` mutates the
buffer in-place, so we were unintentionally mutating the component
binary when reading the buildID. Byte-swap a copy of the buffer so that
the component binary is left untouched.