Convert toolchain version to SDK version (#37)
The toolchain version is a function of the SDK version it targets, but
they are not equivalent. Derive the SDK version from the toolchain
version by stripping off the patch and prerelease fields so that semver
range comparisons with SDK version work as intended.

This fixes the bug where the access_exercise permission is
unintentionally disallowed in toolchain v3.0.0-pre.x versions.